Seperate representations of Boilerplate
